backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
chromadb
google-generativeai
pymupdf
pdfplumber
python-docx
tiktoken
